refactor(frontend): モックの問題文が長すぎて邪魔なのでmock/fixture/へ切り出し
