+++ frontend/src/mocks/fixtures/task2-statement.ts
@@ -1,0 +1,51 @@
+export const DUMMY_TASK2_STATEMENT = String.raw`
+## 問題文
+すずっぴー君はちらし寿司ですか？
+
+**Double Asterisk**
+
+__Double Undeerbar__
+
+*Single Asterisk*
+
+*Single Underbar*
+
+~Single Tilde~
+
+~~Double Tilde~~
+
+
+## 制約
+- aaaa
+- bbbb
+  - bbbb-1
+  - bbbb-2
+- cccc
+  - cccc-1
+    - cccc-1-1
+    - cccc-1-2
+    - cccc-1-3
+  - cccc-2
+  - cccc-3
+
+## 入力
+${"```"}
+$A_{1, 1}$ $\ldots$ $A_{1, W}$
+$A_{2, 1}$ $\ldots$ $A_{2, W}$
+$\vdots$
+$A_{H, 1}$ ... $A_{H, W}$
+${"```"}
+
+## 出力
+察してください。
+
+## h2 ほげほげ
+
+### h3 ほげほげ1
+fuga~~~~~~~~~~~~~~~~
+fuga~~~~~~~~~~~~~~~~
+
+### h3 ほげほげ2
+nya~~~~~~~~~~~~~~~~
+nya~~~~~~~~~~~~~~~~
+`;
